+++ python/GafferSceneUI/CameraTweaksUI.py
@@ -98,7 +98,6 @@ Gaffer.Metadata.registerNode(
 		"tweaks.*" : [
 
 			"tweakPlugValueWidget:allowRemove", True,
-			"tweakPlugValueWidget:allowCreate", True,
 			"tweakPlugValueWidget:propertyType", "parameter",
 
 		],